backend/requirements.txt
flask
flask-cors
boto3
python-dotenv
flask_bcrypt
psycopg2-binary
cryptography